to simplify, offset type is int64_t

+++ modules/imgcodecs/src/grfmt_pam.hpp
@@ -78,7 +78,7 @@ public:
 protected:
 
     RLByteStream    m_strm;
-    size_t m_offset;
+    int64_t m_offset;
     int m_maxval, m_channels, m_sampledepth, selected_fmt;
     bool bit_mode;
 };
